Update v5 names to follow chart name
When deploying the chart, I noticed that lots of the resources
generated do not have a name that can be controlled when doing
'helm install --name <something>'.

I used content generated by 'helm create' + best practices
from other charts and introduced a 'reportportal.fullname' that
is used to build the resource names.